Move user table into page from layout
